refactor: Move Obfuscate interface into api-common-model
