+++ alert-common/src/main/java/com/synopsys/integration/alert/common/rest/model/SettingsProxyModel.java
@@ -12,6 +12,8 @@ import java.util.Optional;
 
 import org.apache.commons.lang3.StringUtils;
 
+import com.synopsys.integration.alert.api.common.model.Obfuscated;
+
 public class SettingsProxyModel extends ConfigWithMetadata implements Obfuscated<SettingsProxyModel> {
     private String proxyHost;
     private Integer proxyPort;
