+++ api-common-model/src/main/java/com/synopsys/integration/alert/api/common/model/Obfuscated.java
@@ -5,7 +5,7 @@
  *
  * Use subject to the terms and conditions of the Synopsys End User Software License and Maintenance Agreement. All rights reserved worldwide.
  */
-package com.synopsys.integration.alert.common.rest.model;
+package com.synopsys.integration.alert.api.common.model;
 
 public interface Obfuscated<T> {
     T obfuscate();
